Optimise scheduler to fetch alert definition only when it's necessary

+++ pkg/services/ngalert/models.go
@@ -17,16 +17,16 @@ type AlertDefinition struct {
 	Updated   int64
 	// Interval in seconds
 	Interval int64
-	Version  int
+	Version  int64
 }
 
 // AlertDefinitionVersion is the model for alert definition versions in Alerting NG.
 type AlertDefinitionVersion struct {
 	ID                int64 `xorm:"pk autoincr 'id'"`
 	AlertDefinitionID int64 `xorm:"alert_definition_id"`
-	ParentVersion     int
-	RestoredFrom      int
-	Version           int
+	ParentVersion     int64
+	RestoredFrom      int64
+	Version           int64
 
 	Created   time.Time
 	Name      string
